Move ljlk test bindings to test module.
Move binding of ljlk potential subcomponents into compiled test module.
Allows potential access in tests for spot checks and potential component
gradchecks. Tidy potential test layout.
